+++ src/render_engine/vulkan/vulkan_render_engine.cpp
@@ -132,104 +132,6 @@ namespace nova {
         return !(*this == other);
     }
 
-    VkShaderModule vulkan_render_engine::create_shader_module(const std::vector<uint32_t>& spirv) const {
-        VkShaderModuleCreateInfo shader_module_create_info;
-        shader_module_create_info.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
-        shader_module_create_info.pNext = nullptr;
-        shader_module_create_info.flags = 0;
-        shader_module_create_info.pCode = spirv.data();
-        shader_module_create_info.codeSize = spirv.size() * 4;
-
-        VkShaderModule module;
-        NOVA_THROW_IF_VK_ERROR(vkCreateShaderModule(device, &shader_module_create_info, nullptr, &module), render_engine_initialization_exception);
-
-        return module;
-    }
-
-    void vulkan_render_engine::upload_new_mesh_parts() {
-        if(mesh_upload_queue.empty()) {
-            // Early out yay
-            return;
-        }
-
-        VkCommandBuffer mesh_upload_cmds;
-
-        VkCommandBufferAllocateInfo alloc_info = {};
-        alloc_info.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-        alloc_info.commandBufferCount = 1;
-        alloc_info.commandPool = get_command_buffer_pool_for_current_thread(copy_family_index);
-        alloc_info.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-
-        vkAllocateCommandBuffers(device, &alloc_info, &mesh_upload_cmds);
-
-        VkCommandBufferBeginInfo begin_info = {};
-        begin_info.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-
-        vkBeginCommandBuffer(mesh_upload_cmds, &begin_info);
-
-        // Ensure that all reads from this buffer have finished. I don't care about writes because the only
-        // way two dudes would be writing to the same region of a megamesh at the same time was if there was a
-        // horrible problem
-
-        mesh_memory->add_barriers_before_data_upload(mesh_upload_cmds);
-
-        mesh_upload_queue_mutex.lock();
-        meshes_mutex.lock();
-
-        std::vector<vk_buffer> freed_buffers;
-        freed_buffers.reserve(mesh_upload_queue.size() * 2);
-        while(!mesh_upload_queue.empty()) {
-            const mesh_staging_buffer_upload_command cmd = mesh_upload_queue.front();
-            mesh_upload_queue.pop();
-
-            compacting_block_allocator::allocation_info* mem = mesh_memory->allocate(cmd.staging_buffer.alloc_info.size);
-
-            VkBufferCopy copy = {};
-            copy.size = cmd.staging_buffer.alloc_info.size;
-            copy.srcOffset = 0;
-            copy.dstOffset = mem->offset;
-            vkCmdCopyBuffer(mesh_upload_cmds, cmd.staging_buffer.buffer, mem->block->get_buffer(), 1, &copy);
-
-            VkDrawIndexedIndirectCommand mesh_draw_command = {};
-            mesh_draw_command.indexCount = (cmd.model_matrix_offset - cmd.indices_offset) / sizeof(uint32_t);
-            mesh_draw_command.instanceCount = 1;
-            mesh_draw_command.firstIndex = 0;
-            mesh_draw_command.vertexOffset = static_cast<uint32_t>(mem->offset);
-            mesh_draw_command.firstInstance = 0;
-
-            meshes[cmd.mesh_id] = {mem, cmd.indices_offset, cmd.model_matrix_offset, mesh_draw_command};
-
-            freed_buffers.insert(freed_buffers.end(), cmd.staging_buffer);
-        }
-
-        mesh_memory->add_barriers_after_data_upload(mesh_upload_cmds);
-
-        mesh_upload_queue_mutex.unlock();
-        meshes_mutex.unlock();
-
-        vkEndCommandBuffer(mesh_upload_cmds);
-
-        VkSubmitInfo submit_info = {};
-        submit_info.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-        submit_info.commandBufferCount = 1;
-        submit_info.pCommandBuffers = &mesh_upload_cmds;
-
-        // Be super duper sure that mesh rendering is done
-        for(const auto& [pass_name, pass] : render_passes) {
-            (void) pass_name;
-            vkWaitForFences(device, 1, &pass.fence, VK_TRUE, std::numeric_limits<uint64_t>::max());
-        }
-        vkQueueSubmit(copy_queue, 1, &submit_info, upload_to_megamesh_buffer_done);
-
-        vkWaitForFences(device, 1, &upload_to_megamesh_buffer_done, VK_TRUE, std::numeric_limits<uint64_t>::max());
-
-        vkResetFences(device, 1, &upload_to_megamesh_buffer_done);
-
-        // Once the upload is done, return all the staging buffers to the pool
-        std::lock_guard l(mesh_staging_buffers_mutex);
-        available_mesh_staging_buffers.insert(available_mesh_staging_buffers.end(), freed_buffers.begin(), freed_buffers.end());
-    }
-
     VKAPI_ATTR VkBool32 VKAPI_CALL debug_report_callback(VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity, VkDebugUtilsMessageTypeFlagsEXT messageTypes, const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData, void* /* pUserData */) {
         std::string type = "General";
         if(messageTypes & VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT) {
@@ -402,150 +304,6 @@ namespace nova {
         }
 
         return VK_FORMAT_R10X6G10X6_UNORM_2PACK16;
-    }
-
-    void vulkan_render_engine::reset_render_finished_semaphores() {
-        render_finished_semaphores_by_frame[current_frame].clear();
-    }
-
-    void vulkan_render_engine::add_resource_to_bindings(std::unordered_map<std::string, vk_resource_binding>& bindings, const spirv_cross::CompilerGLSL& shader_compiler, const spirv_cross::Resource& resource, const VkDescriptorType type) {
-        const uint32_t set = shader_compiler.get_decoration(resource.id, spv::DecorationDescriptorSet);
-        const uint32_t binding = shader_compiler.get_decoration(resource.id, spv::DecorationBinding);
-
-        vk_resource_binding new_binding = {};
-        new_binding.set = set;
-        new_binding.binding = binding;
-        new_binding.descriptorType = type;
-        new_binding.descriptorCount = 1;
-
-        if(bindings.find(resource.name) == bindings.end()) {
-            // Totally new binding!
-            bindings[resource.name] = new_binding;
-
-        } else {
-            // Existing binding. Is it the same as our binding?
-            const vk_resource_binding& existing_binding = bindings.at(resource.name);
-            if(existing_binding != new_binding) {
-                // They have two different bindings with the same name. Not allowed
-                NOVA_LOG(ERROR) << "You have two different uniforms named " << resource.name << " in different shader stages. This is not allowed. Use unique names";
-            }
-        }
-    }
-
-    void vulkan_render_engine::get_shader_module_descriptors(const std::vector<uint32_t>& spirv, std::unordered_map<std::string, vk_resource_binding>& bindings) {
-        const spirv_cross::CompilerGLSL shader_compiler(spirv);
-        const spirv_cross::ShaderResources resources = shader_compiler.get_shader_resources();
-
-        for(const spirv_cross::Resource& resource : resources.sampled_images) {
-            add_resource_to_bindings(bindings, shader_compiler, resource, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER);
-        }
-
-        for(const spirv_cross::Resource& resource : resources.uniform_buffers) {
-            add_resource_to_bindings(bindings, shader_compiler, resource, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER);
-        }
-    }
-
-    std::vector<VkDescriptorSetLayout> vulkan_render_engine::create_descriptor_set_layouts(std::unordered_map<std::string, vk_resource_binding> all_bindings) const {
-        std::unordered_map<uint32_t, std::vector<VkDescriptorSetLayoutBinding>> bindings_by_set;
-
-        for(const auto& named_binding : all_bindings) {
-            const vk_resource_binding& binding = named_binding.second;
-            VkDescriptorSetLayoutBinding new_binding = {};
-            new_binding.binding = binding.binding;
-            new_binding.descriptorCount = binding.descriptorCount;
-            new_binding.descriptorType = binding.descriptorType;
-            new_binding.pImmutableSamplers = binding.pImmutableSamplers;
-            new_binding.stageFlags = VK_SHADER_STAGE_ALL;
-
-            bindings_by_set[binding.set].push_back(new_binding);
-        }
-
-        std::vector<VkDescriptorSetLayoutCreateInfo> dsl_create_infos = {};
-        dsl_create_infos.reserve(bindings_by_set.size());
-        for(uint32_t i = 0; i < bindings_by_set.size(); i++) {
-            if(bindings_by_set.find(i) == bindings_by_set.end()) {
-                NOVA_LOG(ERROR) << "Could not get information for descriptor set " << i << "; most likely you skipped"
-                                << " a descriptor set in your shader. Ensure that all shaders for this pipeline together don't have"
-                                << " any gaps in the descriptor sets they declare";
-                throw shader_layout_creation_failed("Descriptor set " + std::to_string(i) + " not present");
-            }
-
-            const std::vector<VkDescriptorSetLayoutBinding>& bindings = bindings_by_set.at(i);
-
-            VkDescriptorSetLayoutCreateInfo create_info = {};
-            create_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-            create_info.bindingCount = static_cast<uint32_t>(bindings.size());
-            create_info.pBindings = bindings.data();
-
-            dsl_create_infos.push_back(create_info);
-        }
-
-        std::vector<VkDescriptorSetLayout> layouts;
-        layouts.resize(dsl_create_infos.size());
-        for(uint32_t i = 0; i < dsl_create_infos.size(); i++) {
-            vkCreateDescriptorSetLayout(device, &dsl_create_infos[i], nullptr, &layouts[i]);
-        }
-
-        return layouts;
-    }
-
-    std::vector<VkImageMemoryBarrier> vulkan_render_engine::make_attachment_to_shader_read_only_barriers(const std::unordered_set<std::string>& textures) {
-        std::vector<VkImageMemoryBarrier> barriers;
-        barriers.reserve(textures.size());
-
-        for(const std::string& tex_name : textures) {
-            const vk_texture& tex = this->textures.at(tex_name);
-
-            VkImageMemoryBarrier barrier = {};
-            barrier.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER;
-            barrier.srcAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-            barrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT;
-            barrier.newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-            barrier.srcQueueFamilyIndex = graphics_family_index;
-            barrier.dstQueueFamilyIndex = graphics_family_index;
-            barrier.image = tex.image;
-            barrier.subresourceRange.baseMipLevel = 0;
-            barrier.subresourceRange.levelCount = 1;
-            barrier.subresourceRange.baseArrayLayer = 0;
-            barrier.subresourceRange.layerCount = 1;
-
-            if(tex.is_depth_tex) {
-                barrier.oldLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-                barrier.subresourceRange.aspectMask = VK_IMAGE_ASPECT_DEPTH_BIT;
-
-            } else {
-                barrier.oldLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-                barrier.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-            }
-
-            barriers.push_back(barrier);
-        }
-
-        return barriers;
-    }
-
-    void vulkan_render_engine::write_texture_to_descriptor(const vk_texture& texture, VkWriteDescriptorSet& write, std::vector<VkDescriptorImageInfo>& image_infos) const {
-        VkDescriptorImageInfo image_info = {};
-        image_info.imageView = texture.image_view;
-        image_info.imageLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-        image_info.sampler = point_sampler;
-
-        image_infos.push_back(image_info);
-
-        write.pImageInfo = &image_infos.at(image_infos.size() - 1);
-        write.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-    }
-
-    void vulkan_render_engine::write_buffer_to_descriptor(const vk_buffer& buffer, VkWriteDescriptorSet& write, std::vector<VkDescriptorBufferInfo>& buffer_infos) {
-        VkDescriptorBufferInfo buffer_info = {};
-        buffer_info.buffer = buffer.buffer;
-        buffer_info.offset = 0;
-        buffer_info.range = buffer.alloc_info.size;
-
-        buffer_infos.push_back(buffer_info);
-
-        write.pBufferInfo = &buffer_infos[buffer_infos.size() - 1];
-        write.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
     }
 
     void vulkan_render_engine::destroy_render_passes() {
