+++ src/render_engine/vulkan/vulkan_render_engine_mesh.cpp
@@ -3,3 +3,5 @@
  * \date 07-Feb-19.
  */
 
+#include "vulkan_render_engine.hpp"
+
